# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EdgeEver API",
    "version": "0.1.0",
    "description": "REST API for EdgeEver notebooks, memos, resources, API tokens, and Agent integrations."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/users": {
      "get": {
        "summary": "List instance users (owner only)",
        "responses": {
          "200": {
            "description": "Instance user list",
            "content": { "application/json": { "schema": {
              "type": "object",
              "properties": { "users": { "type": "array", "items": { "$ref": "#/components/schemas/InstanceUser" } } },
              "required": ["users"]
            } } }
          },
          "403": { "$ref": "#/components/responses/Error" }
        }
      },
      "post": {
        "summary": "Create an isolated user workspace (owner only)",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserCreateInput" } } } },
        "responses": {
          "201": { "description": "Created user", "content": { "application/json": { "schema": {
            "type": "object", "properties": { "user": { "$ref": "#/components/schemas/InstanceUser" } }, "required": ["user"]
          } } } },
          "403": { "$ref": "#/components/responses/Error" },
          "409": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/v1/users/{id}": {
      "patch": {
        "summary": "Update, disable, or reset a user (owner only)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserUpdateInput" } } } },
        "responses": {
          "200": { "description": "Updated user", "content": { "application/json": { "schema": {
            "type": "object", "properties": { "user": { "$ref": "#/components/schemas/InstanceUser" } }, "required": ["user"]
          } } } },
          "403": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "edgeever_session"
      }
    },
...
    "responses": {
...
      "Error": {
        "description": "API error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {}
            },
            "required": ["code", "message"]
          }
        },
        "required": ["error"]
      },
...
      "AuthUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "displayName": {
            "type": ["string", "null"]
          },
          "role": {
            "type": "string",
            "enum": ["owner", "member"]
          }
        },
        "required": ["id", "username", "displayName", "role"]
      },
      "InstanceUser": {
        "allOf": [
          { "$ref": "#/components/schemas/AuthUser" },
          {
            "type": "object",
            "properties": {
              "isDisabled": { "type": "boolean" },
              "lastLoginAt": { "type": ["string", "null"], "format": "date-time" },
              "createdAt": { "type": "string", "format": "date-time" }
            },
            "required": ["isDisabled", "lastLoginAt", "createdAt"]
          }
        ]
      },
      "UserCreateInput": {
        "type": "object",
        "properties": {
          "username": { "type": "string", "minLength": 1, "maxLength": 80 },
          "displayName": { "type": ["string", "null"], "maxLength": 80 },
          "password": { "type": "string", "minLength": 8, "maxLength": 512 }
        },
        "required": ["username", "password"]
      },
      "UserUpdateInput": {
        "type": "object",
        "properties": {
          "displayName": { "type": ["string", "null"], "maxLength": 80 },
          "password": { "type": "string", "minLength": 8, "maxLength": 512 },
          "isDisabled": { "type": "boolean" }
        },
        "minProperties": 1
      },
...
    }
  }
}